pyproject.toml
[project]
name = "nerdfarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "openai>=2.30.0",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2",
    "streamlit>=1.32.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
